Create CDM/CIM Views
Continued implementing the CDM Interface. Changed the code so that all classes load schema/xml if its available, otherwise
they fall back to building the table in code.
