+++ src/CDMView/CDMView.csproj
@@ -57,12 +57,15 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <None Include="CDMViewSchema.xsd" />
+    <None Include="CDMView.xsd" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="CDMView.xml" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>copy "$(MSBuildProjectDirectory)\bin\$(Configuration)\*.dll" $(SolutionDir)\..\dist
-copy "$(MSBuildProjectDirectory)\CDMViewSchema.xsd" $(SolutionDir)\..\dist
-copy "$(MSBuildProjectDirectory)\CDMViewData.xml" $(SolutionDir)\..\dist</PostBuildEvent>
+copy "$(MSBuildProjectDirectory)\CDMView.xsd" $(SolutionDir)\..\dist
+copy "$(MSBuildProjectDirectory)\CDMView.xml" $(SolutionDir)\..\dist</PostBuildEvent>
   </PropertyGroup>
 </Project>
