+++ src/CDMView/CDMView.xsd
@@ -53,7 +53,6 @@
               <xs:element name="initial" type="xs:string" minOccurs="0" />
               <xs:element name="min" type="xs:string" minOccurs="0" />
               <xs:element name="max" type="xs:string" minOccurs="0" />
-			  <xs:element name="comment" type="xs:string" minOccurs="0" />
             </xs:sequence>
           </xs:complexType>
         </xs:element>
@@ -74,7 +73,22 @@
             </xs:sequence>
           </xs:complexType>
         </xs:element>
+        <xs:element name="Messages">
+          <xs:complexType>
+            <xs:sequence>
+              <xs:element name="type" type="xs:string" />
+              <xs:element name="code" type="xs:string" />
+              <xs:element name="brief" type="xs:string" minOccurs="0" />
+              <xs:element name="description" type="xs:string" minOccurs="0" />
+            </xs:sequence>
+          </xs:complexType>
+        </xs:element>
       </xs:choice>
     </xs:complexType>
+    <xs:unique name="Constraint1" msdata:PrimaryKey="true">
+      <xs:selector xpath=".//Messages" />
+      <xs:field xpath="type" />
+      <xs:field xpath="code" />
+    </xs:unique>
   </xs:element>
 </xs:schema>
