+++ src/CashIn/CashInView.cs
@@ -19,7 +19,9 @@ namespace CashIn
       /// <returns>new HCDU table</returns>
       protected override BaseTable CreateTableInstance(IContext ctx)
       {
-         return new CashInTable(ctx, viewName);
+         CashInTable cashInTable = new CashInTable(ctx, viewName);
+         cashInTable.ReadXmlFile();
+         return cashInTable;
       }
    }
 }
