+++ src/HDCUView/HDCUView.cs
@@ -19,7 +19,9 @@ namespace HDCUView
       /// <returns>new HCDU table</returns>
       protected override BaseTable CreateTableInstance(IContext ctx)
       {
-         return new HDCUTable(ctx, viewName);
+         HDCUTable cduTable = new HDCUTable(ctx, viewName);
+         cduTable.ReadXmlFile();
+         return cduTable;
       }
    }
 }
